+++ controllers/awsmachinepool_controller.go
@@ -18,6 +18,7 @@ package controllers
 
 import (
 	"context"
+	"fmt"
 	"time"
 
 	"github.com/go-logr/logr"
@@ -53,10 +54,21 @@ func (r *AWSMachinePoolReconciler) Reconcile(req ctrl.Request) (ctrl.Result, err
 		logger.Error(err, "AWSMachinePool does not exist")
 		return ctrl.Result{}, err
 	}
+	// check if CR got CAPI watch-filter label
+	if !key.HasCapiWatchLabel(awsMachinePool.Labels) {
+		logger.Info(fmt.Sprintf("AWSMachinePool do not have %s=%s label, ignoring CR", key.ClusterWatchFilterLabel, "capi"))
+		// ignoring this CR
+		return ctrl.Result{}, nil
+	}
 
 	clusterName := key.GetClusterIDFromLabels(awsMachinePool.ObjectMeta)
 
 	logger = logger.WithValues("cluster", clusterName)
+
+	if awsMachinePool.Spec.AWSLaunchTemplate.IamInstanceProfile == "" {
+		logger.Info("AWSMachinePool has empty .Spec.AWSLaunchTemplate.IamInstanceProfile, not creating IAM role")
+		return ctrl.Result{}, nil
+	}
 
 	var awsClientGetter *awsclient.AwsClient
 	{